requirements.txt
python-binance==1.0.16
web3==5.31.1
python-dotenv==0.21.0
numpy==1.24.3
pandas==1.5.3
fastapi==0.95.0
uvicorn==0.21.1
websockets==10.4
aiohttp==3.8.4
pydantic==1.10.7
matplotlib==3.7.1
python-telegram-bot==13.13
dataclasses-json==0.5.7
APScheduler==3.10.1
pytz==2023.3
Pillow==9.5.0
requests==2.28.2
nest_asyncio==1.5.6 